+++ test/framework/servicebinding/servicebinding.go
@@ -10,7 +10,7 @@ import (
 	runtimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/anynines/a8s-deployment/test/framework"
-	"github.com/anynines/a8s-service-binding-controller/api/v1alpha1"
+	"github.com/anynines/a8s-service-binding-controller/api/v1beta3"
 )
 
 const (
@@ -20,14 +20,14 @@ const (
 
 // Option represents a functional option for service binding objects. To learn what a functional
 // option is, read here: https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis
-type Option func(*v1alpha1.ServiceBinding)
+type Option func(*v1beta3.ServiceBinding)
 
 func SetInstanceRef(dsi runtimeclient.Object) Option {
-	return func(sb *v1alpha1.ServiceBinding) {
+	return func(sb *v1beta3.ServiceBinding) {
 		sb.Spec.Instance.APIVersion = dsi.
 			GetObjectKind().GroupVersionKind().GroupVersion().String()
 		sb.Spec.Instance.Kind = dsi.GetObjectKind().GroupVersionKind().Kind
-		sb.Spec.Instance.NamespacedName = v1alpha1.NamespacedName{
+		sb.Spec.Instance.NamespacedName = v1beta3.NamespacedName{
 			Name:      dsi.GetName(),
 			Namespace: dsi.GetNamespace(),
 		}
@@ -37,7 +37,7 @@ func SetInstanceRef(dsi runtimeclient.Object) Option {
 // TODO: Make two separate options for name and namespace. We only need to pass string as
 // parameters
 func SetNamespacedName(dsi runtimeclient.Object) Option {
-	return func(sb *v1alpha1.ServiceBinding) {
+	return func(sb *v1beta3.ServiceBinding) {
 		sb.Name = framework.UniqueName(sbPrefix(dsi.GetName()), suffixLength)
 		sb.Namespace = dsi.GetNamespace()
 	}
@@ -47,8 +47,8 @@ func sbPrefix(dsiName string) string {
 	return fmt.Sprintf("%s-sb", dsiName)
 }
 
-func New(opts ...Option) *v1alpha1.ServiceBinding {
-	sb := &v1alpha1.ServiceBinding{}
+func New(opts ...Option) *v1beta3.ServiceBinding {
+	sb := &v1beta3.ServiceBinding{}
 	for _, opt := range opts {
 		opt(sb)
 	}
@@ -59,7 +59,7 @@ func SecretName(sbName string) string {
 	return fmt.Sprintf("%s-%s", sbName, "service-binding")
 }
 
-func WaitForReadiness(ctx context.Context, sb *v1alpha1.ServiceBinding, c runtimeclient.Client) {
+func WaitForReadiness(ctx context.Context, sb *v1beta3.ServiceBinding, c runtimeclient.Client) {
 	var err error
 	EventuallyWithOffset(1, func() bool {
 		sbCreated := New()